Update user profile,stories and value proposition for v1.1
The target user profile, user stories and value proposition
reflect the requirements of AB3.

This is not representative of the requirements for AIA.

Let's update the target user profile, user stories and
value proposition in linewith designated requirements
of AIA to close AY2122S2-CS2103T-T17-3#3 as part of
the v1.1 increment.