Adjust Reminder field in Person
The current Reminder implementation only allows user to assign at most one Reminder to a contact, which is undesirable as there could be multiple reminders for multiple events with one contact.

To allow multiple Reminder for each contact, we should store a Collection of Reminder instead of a Reminder in the Person class.

Let’s
* Add a ReminderList class for a list that is catered to only the Reminders
* Change the Reminder field in Person to ReminderList

This commit will close AY2122S2-CS2103T-T17-3#49