Add commands to view reminders
The current Reminder system implementation does not provide a way to display reminders. 

This commit will provide two ways to view the reminders. One way is to specify the contact index, in which all of the reminders associated with the specified contact will be displayed. Otherwise users can choose to specify a date, in which all of the reminders happening on the specified date will be displayed. In the event which the date is not given, reminders happening today will be shown instead.

Let's
* Add ListDateReminderCommand to execute list reminders by date command.
* Add ListDateReminderCommandParser to parse user inputs to ListDateReminderCommand.
* Add ListContactReminderCommand to execute list reminders by contact command.
* Add ListContactReminderCommandParser to parse user inputs to ListContactReminderCommand.

This commit will close AY2122S2-CS2103T-T17-3#50