Add backup function
The application will discard existing save file when an error
occurs during reading. This is detrimental because the user will
lose all data if they made a small error during a manual edit of
the save file.

Let's create a backup of the save file upon receipt of an Illegal
Value Exception in the readAddressBook method.

This will close AY2122S2-CS2103T-T17-3#64.