Add support to show all tags
The existing tagging system do not allow user to list all the tags that
are in the AddressBook.

This commit will allow user to list all existing tags in AddressBook.

The tags for the command will not be persistent (not saved in storage),
and will only be kept in memory. During runtime, the list of tags will
be populated in UniqueTagList and then updated accordingly upon every
command affecting the tag list.

Let's,
* Update AddressBook to hold UniqueTagList that will hold all tags
  * Initialise UniqueTagList in the constructor
  * Update all CRUD methods to update UniqueTagList
  * Add methods getUniqueTagList, getNumberOfUniqueTags to handle tags
  * Improve equals' readability
  * Update equals and hashCode to include UniqueTagList in comparison
* Update AddressBookParser
  * Rename isValidCommand to checkEmptyArguments, with method JavaDoc.
  * Add an alternate path in parseCommand to return ListTagsCommand
* Add method signatures for this command to Model
  * Add getUniqueTagList()
  * Add getNumberOfUniqueTags()
  * Update ModelStub in AddCommandTest
* Implement getUniqueTagList(), getNumberOfUniqueTags() in ModelManager
* Update Tag to be comparable alphabetically
* Improve equals' readability in Tag
* Add ListTagsCommand for list tags command execution
* Add interface for UniqueTagList for read-only purposes
* Implement UniqueTagList to keep track of all tags in AddressBook

ReadOnlyUniqueTagList is designed to protect UniqueTagList so that
other classes cannot modify UniqueTagList. It is also reinforced by
getCopy method of UniqueTagList, so that the original copy is immutable
by other classes.

Resolves #33.